sahayakai-main/services/voice-server/pyproject.toml
[project]
name = "sahayakai-voice-server"
version = "0.1.0"
description = "Pipecat-based voice server for SahayakAI Parent Call — API-based providers (no GPU)"
requires-python = ">=3.10"

dependencies = [
    "pipecat-ai[silero,websocket,sarvam]>=0.0.108",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "loguru>=0.7",
]

[project.scripts]
voice-server = "src.main:main"
